# [apiserver] Use existing certs for base tests.