Occasional timeouts with the race build. Smaller keys makes the generation much less CPU intensive.