@howbazaar 2017-01-16
## QA

Running tests with -race
  - go test -c -race
  - compile the test binary once to get that timing out of the calculations

time ./apiserver.test  -gocheck.v -gocheck.f serverSuite
  - went from 44s to 27s, actual test times were very similar, just faster test setup
time ./apiserver.test -gocheck.v -gocheck.f guiSuite
  - went from 56s to 28s
time ./apiserver.test -gocheck.v -gocheck.f guiArchiveSuite
  - went from 29s to 13s